Enhance progress bars and fix empty tracks bug
When I did the list(iterator) trick before to fix
the weird attribute error, it expectedly causes
the iterator to be exhausted by the time the
comprehension is reached.  Because it was empty,
it looked like the code was working, but in
reality, tracks were just being saved as [] i.e.
empty.